+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/worksheet/worksheet_helper.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module OoxmlParser
+  # Module with helper methods for worksheet
+  module WorksheetHelper
+    # @return [True, False] is columns are with default rules
+    def default_columns?
+      return true if @columns.empty?
+      return true unless @columns.first.custom_width
+
+      false
+    end
+  end
+end